# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API service for Arthemis Brain.",
        "title": "Arthemis Brain API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/sdg/": {
            "get": {
                "description": "Retrieves every sdg",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sdg"
                ],
                "summary": "Gets all Sdgs",
                "responses": {
                    "200": {
                        "description": "Sdg deleted successfully",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "404": {
                        "description": "Sdg not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sdg/create": {
            "post": {
                "description": "Creates a new sdg",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sdg"
                ],
                "summary": "Create Sdg",
                "parameters": [
                    {
                        "description": "Sdg details",
                        "name": "sdg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SdgRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Sdg created successfully",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON or bad request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sdg/delete/{id}": {
            "delete": {
                "description": "Deletes a sdg by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sdg"
                ],
                "summary": "Delete Sdg",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sdg ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sdg deleted successfully",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "SdgID not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Sdg not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sdg/update/{id}": {
            "patch": {
                "description": "Updates an existing sdg by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sdg"
                ],
                "summary": "Update Sdg",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sdg ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Sdg details to update",
                        "name": "sdg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SdgRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Sdg"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON or SdgID not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Sdg not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sdg/{id}": {
            "get": {
                "description": "Retrieves a sdg by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sdg"
                ],
                "summary": "Get Sdg",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sdg ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Sdg"
                        }
                    },
                    "400": {
                        "description": "SdgID not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Sdg not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "models.Sdg": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "iconURL": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.SdgRequest": {
            "type": "object",
            "properties": {
                "iconURL": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                }
            }
        },
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Tarefa não encontrada"
                },
                "message": {
                    "type": "string",
                    "example": "ID inválido fornecido"
                }
            }
        }
    }
}